pyproject.toml
[project]
name = "form-assistant"
version = "0.1.0"
description = "Form assistant with MCP integration"
dependencies = [
    "langchain-anthropic>=0.1.0",
    "mcp-use>=1.3.3",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "fastmcp",
    "requests>=2.32.4",
    "jsonschema-pydantic>=0.2.0",
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "mcp"]
